+++ internal/service/hub/handler/nta/node_challenge.go
@@ -11,8 +11,8 @@ import (
 )
 
 var (
-	registerMessage    = "I, %s, am signing this message for registering my intention to operate an RSS3 Node."
-	hideTaxRateMessage = "I, %s, am signing this message for registering my intention to hide the tax rate on Explorer for my RSS3 Node."
+	registrationMessage = "I, %s, am signing this message for registering my intention to operate an RSS3 Node."
+	hideTaxRateMessage  = "I, %s, am signing this message for registering my intention to hide the tax rate on Explorer for my RSS3 Node."
 )
 
 func (n *NTA) GetNodeChallenge(c echo.Context) error {
@@ -30,7 +30,7 @@ func (n *NTA) GetNodeChallenge(c echo.Context) error {
 
 	switch request.Type {
 	case "":
-		data = nta.NodeChallengeResponseData(fmt.Sprintf(registerMessage, strings.ToLower(request.NodeAddress.String())))
+		data = nta.NodeChallengeResponseData(fmt.Sprintf(registrationMessage, strings.ToLower(request.NodeAddress.String())))
 	case "hideTaxRate":
 		data = nta.NodeChallengeResponseData(fmt.Sprintf(hideTaxRateMessage, strings.ToLower(request.NodeAddress.String())))
 	default:
